Refactor FirebaseHandler into its own shared class.
I realized that I didn't have an easy way to share logic between
different ReduxVCR projects, and that there would be a lot of
Firebase-related duplication between /persist and /retrieve.

I created a new 'shared' project. It won't be a package, but it has
enough structure to hold its own dependencies and tests. Other
projects will be able to import from it, but they'll otherwise be
self-contained.

This involved a pretty huge rework.