Add CassetteList story, work on Cassette cosmetics
Previously, the themes were randomly generated on render. This poses
a problem, since it means refreshing the page will change the theme of
the cassette!

Updated the probability function to use a seeded randomization, based
on the cassette's ID. This way, it's consistent.

Also added CassetteList story, and converted Replay to a SFC.